1. unit testing for controller 2. update data type

+++ src/main/java/com/mtech/recycler/entity/Promotion.java
@@ -32,5 +32,5 @@ public class Promotion {
     private Date endDate;
 
     @DynamoDBAttribute
-    private float percentage;
+    private double percentage;
 }
